% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilmFarsi.R
\name{FilmFarsi}
\alias{FilmFarsi}
\title{FilmFarsi}
\usage{
FilmFarsi(name, n)
}
\arguments{
\item{name}{Name of the selected color palette. Currently the choices are \code{}, \code{}, and \code{}.}

\item{n}{Number of the desired colors, from the selected color palette.}
}
\value{
This function returns a vector of colors.
}
\description{
The \code{FilmFarsi} provides a vector of colors, selected from the \code{\link[=FilmFarsiList]{FilmFarsiList}} palettes.
}
\details{
The \code{FilmFarsi} provides a vector of colors.
}
\seealso{
\code{\link[=FilmFarsi]{FilmFarsi}}
}
