% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilmFarsiList.R
\docType{data}
\name{FilmFarsiList}
\alias{FilmFarsiList}
\title{List of FilmFarsi Palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
FilmFarsiList
}
\arguments{
\item{FilmFarsiList}{List of \code{\link[=FilmFarsi]{FilmFarsi}} Palettes.}
}
\value{
This function returns a list of \code{\link[=FilmFarsi]{FilmFarsi}} palettes.
}
\description{
The \code{FilmFarsiList} provides a list of \code{\link[=FilmFarsi]{FilmFarsi}} palettes.
}
\details{
The \code{FilmFarsiList} provides a list of \code{\link[=FilmFarsi]{FilmFarsi}} palettes.
}
\seealso{
\code{\link[=FilmFarsi]{FilmFarsi}}
}
\keyword{datasets}
